# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Get social feed (trending tracks sorted by play count)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Feed tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /feed/{id}/like:
    post:
      operationId: likeFeedItem
      tags: [feed]
      summary: Like or unlike a feed item
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeBody"
      responses:
        "200":
          description: Updated like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedLikeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /connections/search:
    get:
      operationId: searchConnections
      tags: [connections]
      summary: Search registered users by name or university
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
          description: Search query (matched against name and university)
      responses:
        "200":
          description: Matching users with connection status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectionUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connections:
    get:
      operationId: getConnections
      tags: [connections]
      summary: Get user connections
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [discover, friends, requests, sent]
          required: false
      responses:
        "200":
          description: Connection list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectionUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connections/{userId}/connect:
    post:
      operationId: sendConnectionRequest
      tags: [connections]
      summary: Send or withdraw a connection request to a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectBody"
      responses:
        "200":
          description: Connection request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /connections/{userId}/respond:
    post:
      operationId: respondToConnectionRequest
      tags: [connections]
      summary: Accept or decline an incoming connection request
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondBody"
      responses:
        "200":
          description: Response result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespondResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LikeBody:
      type: object
      properties:
        liked:
          type: boolean
...
    FeedLikeResponse:
      type: object
      properties:
        itemId:
          type: string
        likes:
          type: integer
        liked:
          type: boolean
      required:
        - itemId
        - likes
        - liked
...
    Track:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        artistId:
          type: string
        genre:
          type: string
        duration:
          type: string
        durationSeconds:
          type: integer
        coverColor:
          type: string
        audioUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        playCount:
          type: integer
        likes:
          type: integer
        university:
          type: ["string", "null"]
      required:
        - id
        - title
        - artist
        - artistId
        - genre
        - duration
        - durationSeconds
        - coverColor
        - audioUrl
        - coverUrl
        - playCount
        - likes
        - university
...
    ConnectionUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        university:
          type: string
        genre:
          type: string
        coverColor:
          type: string
        avatarUrl:
          type: string
          nullable: true
        mutualCount:
          type: integer
        status:
          type: string
          enum: [none, sent, received, connected]
      required:
        - id
        - name
        - university
        - genre
        - coverColor
        - mutualCount
        - status
...
    ConnectBody:
      type: object
      properties:
        connect:
          type: boolean
    ConnectResponse:
      type: object
      properties:
        userId:
          type: string
        status:
          type: string
          enum: [none, sent, received, connected]
      required:
        - userId
        - status
    RespondBody:
      type: object
      properties:
        accept:
          type: boolean
      required:
        - accept
    RespondResponse:
      type: object
      properties:
        userId:
          type: string
        status:
          type: string
          enum: [none, sent, received, connected]
      required:
        - userId
        - status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...